cluster: split the epoch file-lock concern out of heartbeat_epoch.go
The modularity gate caught heartbeat_epoch.go crossing the 1500 LOC
[WATCH] floor (1465 -> 1557) on this branch. Recording an accepted
exception was the other option and would have been the wrong one: the
addition is not entangled with what the file is about.

The acquisition policy is a self-contained concern — how long to wait
for a lock another PROCESS holds, and which errno means "wait" rather
than "give up" — and it is precisely the part a reader chasing a startup
delay needs, without the epoch chaining and wire-format reasoning around
it. withEpochFileLock, acquireEpochFileLock and their two constants move
together to heartbeat_epoch_lock.go.

Pure move: 1359 + 215 lines, no behaviour change, the #6826 cells pass
unchanged.
